fix: two signed overflows on the -w/-y and -x paths
Both found with UBSan, both reachable from a legal command line.

The first is mine, from the granted-window reporting added last week:

  slowhttptest.cc:672: signed integer overflow: 2000000000 * 4
                       cannot be represented in type 'int'

-w and -y go through parse_int with its default bound of INT_MAX, so the
comparison against four times the request overflows well inside the accepted
range. The wrapped negative also made the comparison true, so the warning fired
on a kernel that had clamped the request DOWN to maxsockbuf -- reporting an
override where the opposite had happened. Widened to long long; the same shape
in ng is widened too, where the 1048576 cap on -w/-y means it could not
overflow but the bound and the arithmetic live in different files.

The second is older and worse, since it is a crash rather than UB alone:

  slowhttptest.cc:288: signed integer overflow: 2000000000 * 2
  libc++abi: terminating due to uncaught exception of type std::length_error

-x is accepted up to INT_MAX, doubled into extra_data_max_len_total_, and then
handed to a string that cannot be that long. -x 2000000000 takes the process
down. It is now bounded at 65536: a followup name/value pair has no business
being larger than a header, every other size option here is capped in that
range, and a refusal beats a crash. ng accepts the same value without dying, so
this one is classic-only.

Verified after the change: the huge values are refused with a message, legal
values still run, and the CI smoke test plus both over-long-value refusals pass.

+++ src/slowhttptest.cc
@@ -669,7 +669,12 @@ void SlowHTTPTest::report_status(bool to_stats) {
                cLGN "receive window:" cLGN "      %d requested, %d granted\n",
                requested_window_, granted_window_);
       window_note = line;
-      if(granted_window_ > requested_window_ * 4) {
+      // Widened: -w/-y accept up to INT_MAX here, so the multiply overflows a
+      // signed int well inside the accepted range. UBSan catches it at
+      // -w 2000000000, and the wrapped negative made the comparison true, so
+      // the warning also fired on a kernel that had clamped the request DOWN
+      // -- the opposite of what it reports.
+      if(granted_window_ > static_cast<long long>(requested_window_) * 4) {
         snprintf(line, sizeof(line),
                  cLRD "  WARNING: kernel granted %dx the request; -w/-y are not"
                  " in force\n",
